requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.1
django-filter==1.0.3
django-filters==0.2.1
django-import-export==0.5.1
django-markdown==0.8.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
et-xmlfile==1.0.1
idna==2.5
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
kombu==4.1.0
Markdown==2.6.8
MarkupSafe==1.0
odfpy==1.3.5
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.8
Pillow==4.2.1
python-decouple==3.0
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.2
simplejson==3.11.1
tablib==0.11.5
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
xlrd==1.0.0
xlwt==1.2.0
psycopg2
gunicorn